Building components with Babel
Adding some devtool to build published components. Babel to compile the
components for general consumption and eslint with a config to lint
while editing.

I started to take a crack at how to integrate a theme using FillSpan as
an example. I think there's a lot of work to do there, but wanted to get
this published for now.